make index.js support multiple models (still need to update the client and the evaluation software)
