+++ server/index.js
@@ -4,8 +4,9 @@ import { fileURLToPath } from 'url';
 import cors from 'cors';
 import bodyParser from 'body-parser';
 
-import { GameBoard } from '../common/game.mjs';
-import { minimax, heuristic } from './minimax.js';
+import { GameBoard } from 'common/game';
+import { minimax } from './minimax.js';
+import { heuristic } from './heuristic.js';
 // Define __dirname using ES module syntax
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
@@ -31,12 +32,38 @@ app.get('/', (req, res) => res.redirect('/client'));
 app.use('/client', express.static(path.join(__dirname, '../client')));
 app.use('/common', express.static(path.join(__dirname, '../common')));
 
-app.get('/api/next-move/:xoformat', (req, res) => {
-    const gameBoard = new GameBoard(req.params.xoformat);
-    console.log(`${req.params.xoformat}; heuristic: ${heuristic(gameBoard, 'X')}`);
-    const bestCol = minimax(gameBoard); 
-    console.log(`Received game state ${req.params.xoformat}. The AI chooses to move in column index ${bestCol}`);
-    res.send(String(bestCol)); // Send column as a response. Use setTimeout to simulate it taking a full second.
+app.get('/api/next-move/:model/:xoformat', (req, res) => {
+    const model = req.params.xoformat;
+    const xoformat = req.params.xoformat;
+    const gameBoard = new GameBoard(xoformat);
+    if (gameBoard.status() !== 'U') {
+        const message = "Game is complete. No move can be made: " + xoformat;
+        console.error(message);
+        res.status(400).send(message);
+        return;
+    }
+    console.log(`${xoformat}; heuristic: ${heuristic(gameBoard, 'X')}`);
+    let bestCol;
+    switch (model) {
+        case "random":
+            const availableCols = gameBoard.availableColumns()
+            bestCol = availableCols[Math.floor(Math.random() * availableCols.length)];
+        case "minimax":
+            // evaluate minimax with no heuristic (heuristic function = 0)
+            bestCol = minimax(gameBoard, (gameBoard, nextPlayer) => 0);
+            break;
+        case "minimax-with-heuristic":
+            // evaluate minimax with heuristic
+            bestCol = minimax(gameBoard, heuristic); 
+            break;
+        default:
+            const message = `Model '${model}' not found`;
+            console.error(message);
+            res.status(400).send(message);
+            return;
+    }
+    console.log(`AI chooses to move in column index ${bestCol} for the received game state: ${xoformat}. `);
+    res.send(String(bestCol)); // send column as a response
 });
 
 app.listen(port, () => console.log(`Listening on port ${port}`));
